Centralize and reorg API code.
