+++ lib/traceview.rb
@@ -40,8 +40,7 @@ begin
 
   require 'traceview/config'
   require 'traceview/loading'
-  require 'traceview/legacy_method_profiling'
-  require 'traceview/method_profiling'
+  require 'traceview/api'
 
   if TraceView.loaded
     require 'traceview/instrumentation'
